refactor: flatten secret: and config: commands into get/list/set
Replace `secret:get`, `secret:list`, `config:get`, `config:list` with
top-level `get`, `list`, `set` commands. All commands now use `--env`
as a required flag instead of a positional argument.

- `get --env <env> <path>`: resolves value, fetches from provider if
  SecretRef, prints plain scalars, errors on subtrees
- `list --env <env>`: lists all leaf paths with (secret) suffix
- `set --env <env> <path> [value]`: renamed from secret:set, same behavior
